+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## Unreleased
+
+- feat: Allow closing of SDK (#467)
+
 ## 3.0.7
 
-- fix: export map and sideEffects
+- fix: export map and sideEffects (#464)
 - fix: Don't capture window titles by default (#463)
 - fix: Don't throw on HTTP errors (#458)
 - fix: Throw error if main process code is loaded in renderer (#457)